+++ app/Http/Controllers/Api/OfferController.php
@@ -19,6 +19,7 @@ use Illuminate\Support\Facades\Log;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
 
+
 class OfferController extends Controller
 {
     /**
@@ -76,7 +77,6 @@ class OfferController extends Controller
     public function store(OfferStoreRequest $request)
     {
         $validatedData = $this->validateRedundantInput( $request->validated() );
-
         $offer = Offer::create($validatedData);
         $offer->save();
         $this->saveRelatedData( $offer, $validatedData );
@@ -125,6 +125,7 @@ class OfferController extends Controller
      */
     public function updateExternal(OfferExternalUpdateRequest $request, Institution $institution, String $externalId)
     {
+
         $validatedData = $this->validateRedundantInput( $request->validated() );
 
         $offer = Offer::where(["institution_id" => $institution->id, "externalId" => $externalId ])->first();